data-collection/requirements.txt
numpy
pyyaml
tqdm
stem
selenium
tbselenium
dpkt
gdown
pytest
